+++ MoneyFi-backend/MoneyFi-expense/src/main/java/com/moneyfi/expense/api/ExpenseApiController.java
@@ -30,7 +30,7 @@ public class ExpenseApiController {
     @Operation(summary = "Method to add the expense transaction")
     @PostMapping("/{userId}")
     public ResponseEntity<ExpenseModel> saveExpense(@RequestBody ExpenseModel expense,
-                                                    @PathVariable("userId") int userId) {
+                                                    @PathVariable("userId") Long userId) {
         expense.setUserId(userId);
         ExpenseModel createdExpense = expenseService.save(expense);
         if (createdExpense != null) {
@@ -42,14 +42,14 @@ public class ExpenseApiController {
 
     @Operation(summary = "Method to get all the expenses of a particular user")
     @GetMapping("/{userId}")
-    public ResponseEntity<List<ExpenseModel>> getAllExpenses(@PathVariable("userId") int userId) {
+    public ResponseEntity<List<ExpenseModel>> getAllExpenses(@PathVariable("userId") Long userId) {
         List<ExpenseModel> list = expenseService.getAllexpenses(userId);
-            return ResponseEntity.status(HttpStatus.OK).body(list); // 200
+        return ResponseEntity.status(HttpStatus.OK).body(list); // 200
     }
 
     @Operation(summary = "Method to get all the expense details in a particular month and in a particular year")
     @GetMapping("/{userId}/{month}/{year}/{deleteStatus}")
-    public ResponseEntity<List<ExpenseModel>> getAllExpensesByDate(@PathVariable("userId") int userId,
+    public ResponseEntity<List<ExpenseModel>> getAllExpensesByDate(@PathVariable("userId") Long userId,
                                                                    @PathVariable("month") int month,
                                                                    @PathVariable("year") int year,
                                                                    @PathVariable("deleteStatus") boolean deleteStatus){
@@ -58,9 +58,9 @@ public class ExpenseApiController {
 
     @Operation(summary = "Method to generate the Excel report for the Monthly expenses of a user")
     @GetMapping("/{userId}/{month}/{year}/generateMonthlyReport")
-    public ResponseEntity<byte[]> getMonthlyExpenseReport(@PathVariable("userId") int userId,
-                                                         @PathVariable("month") int month,
-                                                         @PathVariable("year") int year) throws IOException {
+    public ResponseEntity<byte[]> getMonthlyExpenseReport(@PathVariable("userId") Long userId,
+                                                          @PathVariable("month") int month,
+                                                          @PathVariable("year") int year) throws IOException {
 
         byte[] excelData = expenseService.generateMonthlyExcelReport(userId, month, year);
 
@@ -72,7 +72,7 @@ public class ExpenseApiController {
 
     @Operation(summary = "Method to get all the expense details in a particular year")
     @GetMapping("/{userId}/{year}/{deleteStatus}")
-    public ResponseEntity<List<ExpenseModel>> getAllExpensesByYear(@PathVariable("userId") int userId,
+    public ResponseEntity<List<ExpenseModel>> getAllExpensesByYear(@PathVariable("userId") Long userId,
                                                                    @PathVariable("year") int year,
                                                                    @PathVariable("deleteStatus") boolean deleteStatus){
         return ResponseEntity.status(HttpStatus.OK).body(expenseService.getAllexpensesByYear(userId, year, deleteStatus));
@@ -80,8 +80,8 @@ public class ExpenseApiController {
 
     @Operation(summary = "Method to generate the Excel report for the Yearly Expenses of a user")
     @GetMapping("/{userId}/{year}/generateYearlyReport")
-    public ResponseEntity<byte[]> getYearlyExpenseReport(@PathVariable("userId") int userId,
-                                                        @PathVariable("year") int year) throws IOException {
+    public ResponseEntity<byte[]> getYearlyExpenseReport(@PathVariable("userId") Long userId,
+                                                         @PathVariable("year") int year) throws IOException {
 
         byte[] excelData = expenseService.generateYearlyExcelReport(userId, year);
 
@@ -93,13 +93,13 @@ public class ExpenseApiController {
 
     @Operation(summary = "Method to get the total expense amount in a particular month and in a particular year")
     @GetMapping("/{userId}/totalExpense/{month}/{year}")
-    public Double getTotalExpenseByMonthAndYear(@PathVariable("userId") int userId, @PathVariable("month") int month, @PathVariable("year") int year){
+    public Double getTotalExpenseByMonthAndYear(@PathVariable("userId") Long userId, @PathVariable("month") int month, @PathVariable("year") int year){
         return expenseService.getTotalExpenseInMonthAndYear(userId, month, year);
     }
 
     @Operation(summary = "Method to get the list of total expense amount of all months in a particular year")
     @GetMapping("/{userId}/monthlyTotalExpensesList/{year}")
-    public List<Double> getMonthlyTotals(@PathVariable("userId") int userId,
+    public List<Double> getMonthlyTotals(@PathVariable("userId") Long userId,
                                          @PathVariable("year") int year) {
         return expenseService.getMonthlyExpenses(userId, year);
     }
@@ -107,7 +107,7 @@ public class ExpenseApiController {
     @Operation(summary = "Method to get the total expense amount up to previous month (excludes current month)")
     @GetMapping("/{userId}/totalExpensesUpToPreviousMonth/{month}/{year}")
     public Double getTotalExpensesUpToPreviousMonth(
-            @PathVariable("userId") int userId,
+            @PathVariable("userId") Long userId,
             @PathVariable("month") int month,
             @PathVariable("year") int year) {
 
@@ -116,7 +116,7 @@ public class ExpenseApiController {
 
     @Operation(summary = "Method to find the total savings/ remaining amount in a month")
     @GetMapping("/{userId}/totalSavings/{month}/{year}")
-    public Double getTotalSavingsByMonthAndDate(@PathVariable("userId") int userId,
+    public Double getTotalSavingsByMonthAndDate(@PathVariable("userId") Long userId,
                                                 @PathVariable("month") int month,
                                                 @PathVariable("year") int year){
 
@@ -125,14 +125,14 @@ public class ExpenseApiController {
 
     @Operation(summary = "Method to get the list of cumulative savings monthly wise in a year")
     @GetMapping("/{userId}/monthlyCumulativeSavingsInYear/{year}")
-    public List<Double> getCumulativeMonthlySavings(@PathVariable("userId") int userId, @PathVariable("year") int year){
+    public List<Double> getCumulativeMonthlySavings(@PathVariable("userId") Long userId, @PathVariable("year") int year){
 
         return expenseService.getCumulativeMonthlySavings(userId, year);
     }
 
     @Operation(summary = "Method to update the expense details")
     @PutMapping("/{id}")
-    public ResponseEntity<ExpenseModel> updateExpense(@PathVariable("id") int id,
+    public ResponseEntity<ExpenseModel> updateExpense(@PathVariable("id") Long id,
                                                       @RequestBody ExpenseModel expense) {
         ExpenseModel expenseModel = expenseRepository.findById(id).orElse(null);
         if(expenseModel != null){
@@ -155,7 +155,7 @@ public class ExpenseApiController {
 
     @Operation(summary = "Method to delete the particular expense. Here which is typically soft delete")
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteExpenseById(@PathVariable("id") int id) {
+    public ResponseEntity<Void> deleteExpenseById(@PathVariable("id") Long id) {
         boolean isDeleted = expenseService.deleteExpenseById(id);
         if (isDeleted) {
             return ResponseEntity.status(HttpStatus.NO_CONTENT).build(); // 204: No Content
